requirements.txt
Flask
flask-cors
pandas
joblib
scikit-learn==1.6.1
google-generativeai
python-dotenv
gunicorn
lightgbm
gspread
google-auth
